requirements.txt
grpcio==1.62.0
grpcio-tools==1.60.1
googleapis-common-protos==1.62.0
grpcio-status==1.62.0
protobuf==4.25.3
hvac==2.1.0
prometheus_client==0.20.0
pydantic-settings==2.2.1
pydantic==2.7.0
